Factor out frontend.initial_props, .static_content.
