+++ norent/letter_sending.py
@@ -10,7 +10,7 @@ from project import slack, locales
 from project.util.email_attachment import email_file_response_as_attachment
 from project.util.html_to_text import html_to_text
 from project.util.site_util import get_site_of_type, SITE_CHOICES
-from frontend.views import render_raw_lambda_static_content
+from frontend.static_content import render_raw_lambda_static_content
 from frontend.lambda_response import LambdaResponse
 from users.models import JustfixUser
 from loc.views import render_pdf_bytes
